ksmbd-tools: respect sysconfdir
The programs did not respect sysconfdir and expected ksmbdpwd.db and
smb.conf to be under the /etc/ksmbd directory by default. Modify the
autotools and meson builds such that sysconfdir is respected. In the
autotools build, AM_CPPFLAGS is the correct place to put preprocessor
flags, but seeing how AM_CFLAGS is already used, use the latter.

Signed-off-by: atheik <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>