ksmbd-tools: simplify argument handling of ksmbd.mountd
Simplify argument handling of ksmbd.mountd such that it is equivalent
to the other programs. Since the flag member of option is always NULL,
getopt_long() won't return 0. It won't return 1 either since its
opstring argument does not start with '-'. Remove the custom error
message for missing arguments. While we're at it, fix wrongly indented
closing braces in the argument handling sections of the other programs.

Signed-off-by: atheik <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>