ksmbd-tools: replace setsid with setpgid in nodetach mode
In nodetach mode, setsid() was used by the manager process to become a
session leader. However, according to the nearby comment, the intent
is to only become a process group leader. Presumably this is done to
prevent SIGINT and SIGHUP at the terminal from propagating to the
manager process. Be more explicit about this intent and replace
setuid() with setpgid(). Change the wording slightly in the related
usage string. Also, remove the leftover mention of the systemd service
mode which was removed.

Signed-off-by: atheik <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>