ksmbd-tools: move lock file to runstatedir
Since /tmp (or rather TMPDIR) can be subject to arbitrary sweeps, it is
not a good place to store pid files and other per-process temporary
files. The proper place for these files would be the historic /var/run
or the modern /run. In autoconf 2.70, this directory for modifiable
per-process data is configurable using the --runstatedir option. Many
distributions backport this option to earlier versions of autoconf.
Introduce the --with-rundir and -Drundir options to the autotools and
meson builds, respectively. Allow also the use of the --runstatedir
option in autoconf if it is supported. Since meson will hopefully have
a builtin --runstatedir option in the future, make accommodations for
that as well. The project specific --with-rundir or -Drundir option has
precedence over the --runstatedir option. Move the location of the
manager pid file, referred to as the lock file in the project, to
runstatedir. Update the README and systemd unit file template regarding
this change. Also, do leftover whitespace cleanup in automake files.

Signed-off-by: atheik <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>