ksmbd-tools: remove getuid() check from control
There is no reason why control in particular should always be run as
root. Remove the getuid() check. The errno derived error messages
"Operation not permitted" and "Permission denied" give obvious reasons
to the user as to what went wrong. While we're at it, add an error
message if ksmbd failed to be shut down like the one with mountd. Only
return EXIT_SUCCESS if both mountd and ksmbd were shut down.

Signed-off-by: atheik <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>