ksmbd-tools: add man pages
Add man pages for the utilities, the configuration file, and the user
database. Generate the man pages from templates so as to include the
build-time configuration. With the introduction of man pages, make the
usage strings of the utilities less verbose but try still to be exact.
Rewrite the share parameter descriptions in order to better describe
what is implemented. Refer to `shares' instead of `services' and
`parameters' instead of `options'. In the example `smb.conf', set all
global and share parameters to their default values and change the
example share name to `example'. Document the quirk regarding
user-input passwords being always assumed to be UTF-8 encoded.

Signed-off-by: atheik <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>