ksmbd-tools: add @group support to ksmbd.conf for lists of users
For parameters that accept a list of users, make it possible to specify
all users belonging to a system group by prefixing the group name with
`@'. Add the `grc' parameter to parse_list(), which is the character
prefix that identifies a group name. Group members are fetched with
getgrnam(), which returns them in a format identical to the `list'
parameter of parse_list(), i.e. a NULL terminated list of strings.
Since the strings returned by getgrnam() should not be freed, being
available only until a subsequent getgrnam() call is made, make it so
that the caller is responsible for freeing `list'. It is then possible
to simply call parse_list() again in parse_list() to parse the group
members. When `grc' is the null terminator, group member lookup is
disabled. Note, that it is not possible to confuse an empty string for
the disabled prefix due to the preceding cp_ltrim() condition.

Co-authored-by: Wang Junjie <iamwudy@qq.com>
Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>