ksmbd-tools: build utilities as a single binary
Rather than have four different binaries, i.e. ksmbd.addshare,
ksmbd.adduser, ksmbd.control, and ksmbd.mountd, each one including
its own libksmbdtools.a copy, build them instead as a single binary.
This resulting ksmbd.tools acts like BusyBox in that its behavior
depends on the name by which it is called. Then, each utility is made
into a symlink to it, meaning that users keep using the utilities as
usual. ksmbd.tools itself is installed to libexecdir as it is not
something the user should run.

Instead of libksmbdtools.a, each utility becomes its own static
library, i.e. libaddshare.a, libadduser.a, libcontrol.a, libmountd.a,
which are all linked to the single binary. Note that the single binary
approach is also beneficial when statically linking to any of the
external dependencies, e.g. GLib, as it greatly reduces the overall
binary size when there is overlap in copies otherwise made separately
for multiple utilities.

Due to install_symlink(), minimum meson version is bumped to 0.61.5,
meaning that it has to be installed using `pip' in Travis, since no
Ubuntu release currently packages that version or newer. However, bump
to Ubuntu Jammy anyways, just for the sake of building against newer
versions of ksmbd-tools' dependencies.

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>