ksmbd-tools: validate --option in addshare
The purpose of addshare is to provide an interface for configuring
ksmbd as it is running. This means that changing global parameters is
not possible. However, we can still edit the global section as long as
we only change share parameters within it, i.e. change their defaults
for all shares. With `--option', validate that only share parameters
are accepted and allow changes to the global section.

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>