ksmbd-tools: improve mountd daemonization and signal handling
Currently, when mountd daemonizes, the PID file is *not* created before
the mountd manager parent exits. This is a problem for service managers
since they assume that the mountd manager parent exiting with status 0
means that the mountd manager PID is in the PID file. This behavior is
not possible with daemon(3). Drop it and do a service-manager-friendly
daemonization. The mountd manager signals successful daemonization by
sending the parent SIGUSR1. Note that we still do *not* perform the
double-fork trick for never acquiring a controlling terminal.

Currently, mountd also has problems regarding signal handling. Namely,
the handler mostly calls functions which are not async-signal-safe.
Also, the signal-block and signal-ignore masks are inherited from the
parent and are not reset. There is also the issue of the thread signal
mask as the helper thread spawned by g_thread_pool_new() inherits it
and passes it down to threads spawned with g_thread_pool_push().
Address these issues by only calling async-signal-safe functions in the
handler, resetting the signal-block and signal-ignore masks, and
blocking the signals before spawning the helper thread, respectively.
Also, block the signals when accessing global data accessed also in the
handler, so if the need arises, *some* not async-signal-safe functions
may be called.

Note that using pthread_sigmask() over sigprocmask() is a matter of
picking your poison, the use of the former assumes that GLib threads
are POSIX threads (sane) while the use of the latter assumes that it
only modifies the thread signal mask (sane).

Also, note that we now use select() in ipc_process_event() so a config
reload takes effect *immediately* rather than at the next IPC event.
We remove the KSMBD_SHOULD_RELOAD_CONFIG flag only on success.

Finally, note that the use of sigaction() with SA_NOCLDWAIT assumes
that the child still sends SIGCHLD to the parent, which is a Linuxism.
Alternatively, each sigwaitinfo() for SIGCHLD could be followed by a
waitpid() so as to prevent zombies.

With the above changes, an abort() in a GThreadPool thread (e.g. due to
GLib OoM) no longer results in a variety of use-after-frees. Also,
neither a segfault nor an initial pwddb/smbconf parse fail result in a
worker process restart.

As for other changes, don't zero `global_conf' with memset() since it
it is already zeroed and we do not depend on the padding being zero.

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>